+++ packages/core/src/languages/css/html-ids.ts
@@ -1,20 +1,28 @@
 import ManglerExpression from "../utils/mangler-expression.class";
 
 const expressions: ManglerExpression[] = [
-  // e.g.
-  //  #(foo)
-  //  #(foo)({)
-  //  #(foo)( ){
-  //  #(foo)(,) .bar {
-  //  #(foo)(.)bar {
-  //  #(foo)([)data-value] {
-  //  #(foo)(:)focus {
-  //  #(foo)(:):before {
-  //  .bar:not\(#(foo)(\)) {
+  // ID selectors, e.g.:
+  //  `(#foo){ }`
+  //  `(#foo) { }`
+  //  `(#foo), .bar { }`
+  //  `(#foo).bar { }`
+  //  `(#foo)#no-match { }`
+  //  `(#foo)[data-value] { }`
+  //  `(#foo):focus { }`
+  //  `(#foo)::before { }`
+  //  `div(#foo) { }`
+  //  `div:not\((#foo)\) { }`
+  //  `(#foo)>div { }`
+  //  `(#foo)+div { }`
+  //  `(#foo)~div { }`
+  //  `div>(#foo) { }`
+  //  `div+(#foo) { }`
+  //  `div~(#foo) { }`
+  //  `div { } (#foo) { }`
   new ManglerExpression(
-    "\\#(%s)([{\\s,\\.\\[\\:\\)])",
+    "\\#(%s)(?=\\{|\\,|\\.|\\#|\\[|\\:|\\)|\\>|\\+|\\~|\\s)",
     ManglerExpression.matchParserForIndex(1),
-    ManglerExpression.matchReplacerBy("#%s$2"),
+    ManglerExpression.matchReplacerBy("#%s"),
   ),
 ];